requirements/local.txt
-r base.txt
-e .
-e ../base
-e ../block
-e ../login
-e ../mail
beautifulsoup4
devpi-client
django-debug-toolbar
django_extensions
fabric
factory-boy
gitpython
hgapi
ipdb
psycopg2
pytest-django
pytest-flakes
pytest-pep8
pyyaml
walkdir
